requirements.txt
crewai==0.130.0
fastapi==0.110.3
uvicorn==0.29.0
python-multipart==0.0.9
langchain-openai==0.1.7
PyPDF2==3.0.1
celery==5.4.0
redis==5.0.4
sqlalchemy==2.0.30
python-dotenv==1.0.1
requests==2.32.3